+++ src/subscriptions/v2/subscriptions-v2.test.ts
@@ -17,6 +17,7 @@ const licenseThreeUuid = uuidv4();
 const perSeatBasicLicenseUuids = [uuidv4(), uuidv4(), uuidv4(), uuidv4(), uuidv4(), uuidv4()];
 const testGrantee = '123456';
 const testEmail = 'tester@domain.com';
+const owner = 'subscription-owner'
 
 describe('Subscriptions V2 Tests', () => {
   const apiKey = testUuids.devApiKeyV2;
@@ -38,7 +39,7 @@ describe('Subscriptions V2 Tests', () => {
       owner: 'example',
       granteeId: 'test-grantee-id',
       status: 'ACTIVE',
-      endTime: '2025-07-06T12:00:00.000Z',
+      expiryDate: '2025-07-06T12:00:00.000Z',
     });
 
     expect(data).toEqual(expect.objectContaining(subscriptionSchema));
@@ -102,6 +103,20 @@ describe('Subscriptions V2 Tests', () => {
         }),
       ]),
     );
+  });
+
+  it('getAll (w/ search params owner): Should successfully fetch subscriptions', async () => {
+    const dataWithSearchParams = await salable.subscriptions.getAll({
+      owner: 'different-owner'
+    });
+
+    expect(dataWithSearchParams).toEqual({
+      first: expect.any(String),
+      last: expect.any(String),
+      data: expect.arrayContaining([{ ...subscriptionSchema }]),
+    });
+    expect(dataWithSearchParams.data.length).toEqual(1);
+    expect(dataWithSearchParams.data).toEqual([{...subscriptionSchema, owner: 'different-owner'}]);
   });
 
   it('getOne: Should successfully fetch the specified subscription', async () => {
@@ -187,6 +202,14 @@ describe('Subscriptions V2 Tests', () => {
     expect(data).toEqual(expect.objectContaining({ eventUuid: expect.any(String) }));
   });
 
+  it('update: Should successfully update a subscription owner', async () => {
+    const data = await salable.subscriptions.update(perSeatSubscriptionUuid, {
+      owner: 'updated-owner',
+    });
+
+    expect(data).toEqual({ ...subscriptionSchema, owner: 'updated-owner' });
+  });
+
   it('cancel: Should successfully cancel the subscription', async () => {
     const data = await salable.subscriptions.cancel(subscriptionUuid, { when: 'now' });
 
@@ -234,6 +257,7 @@ const subscriptionSchema: Subscription = {
   isTest: expect.any(Boolean),
   cancelAtPeriodEnd: expect.any(Boolean),
   email: expect.toBeOneOf([expect.any(String), null]),
+  owner: expect.toBeOneOf([expect.any(String), null]),
   organisation: expect.any(String),
   quantity: expect.any(Number),
   status: expect.any(String),
@@ -509,6 +533,7 @@ const generateTestData = async () => {
       paymentIntegrationSubscriptionId: stripeEnvs.basicSubscriptionTwoId,
       lineItemIds: [stripeEnvs.basicSubscriptionTwoLineItemId],
       email: testEmail,
+      owner,
       type: 'salable',
       status: 'ACTIVE',
       organisation: testUuids.organisationId,
@@ -527,6 +552,7 @@ const generateTestData = async () => {
       paymentIntegrationSubscriptionId: stripeEnvs.basicSubscriptionThreeId,
       lineItemIds: [stripeEnvs.basicSubscriptionThreeLineItemId],
       email: testEmail,
+      owner,
       type: 'salable',
       status: 'ACTIVE',
       organisation: testUuids.organisationId,
@@ -545,6 +571,7 @@ const generateTestData = async () => {
       paymentIntegrationSubscriptionId: stripeEnvs.proSubscriptionId,
       lineItemIds: [stripeEnvs.proSubscriptionLineItemId],
       email: testEmail,
+      owner: 'different-owner',
       type: 'salable',
       status: 'ACTIVE',
       organisation: testUuids.organisationId,
@@ -563,6 +590,7 @@ const generateTestData = async () => {
       paymentIntegrationSubscriptionId: stripeEnvs.perSeatBasicSubscriptionId,
       uuid: perSeatSubscriptionUuid,
       email: testEmail,
+      owner,
       type: 'salable',
       status: 'ACTIVE',
       organisation: testUuids.organisationId,